Don't reinit db on deploy
At the moment we try to initialize the db on every deploy.
This is the default setting in the parent helm chart
https://github.com/keitaroinc/ckan-helm/blob/master/values.yaml#L86

This is causing our build to fail https://github.com/nhsx/standards-registry/runs/4251391439?check_suite_focus=true#step:7:17
as the db already exists

For now: just use the existing db on deploy. We can split into an init
and an update job in the future